+++ src/Classifiers/NaiveBayes.php
@@ -11,6 +11,7 @@ use Rubix\ML\Probabilistic;
 use Rubix\ML\EstimatorType;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Other\Helpers\Params;
+use Rubix\ML\Other\Traits\TracksRevisions;
 use Rubix\ML\Specifications\DatasetIsLabeled;
 use Rubix\ML\Specifications\DatasetIsNotEmpty;
 use Rubix\ML\Specifications\SpecificationChain;
@@ -43,6 +44,8 @@ use const Rubix\ML\LOG_EPSILON;
  */
 class NaiveBayes implements Estimator, Learner, Online, Probabilistic, Persistable
 {
+    use TracksRevisions;
+
     /**
      * The amount of (Laplace) smoothing added to the probabilities.
      *