+++ src/Persisters/Serializers/Igbinary.php
@@ -7,14 +7,13 @@ use Rubix\ML\Persistable;
 use Rubix\ML\Exceptions\RuntimeException;
 use __PHP_Incomplete_Class;
 
-use function is_null;
+use function extension_loaded;
 use function is_object;
 
 /**
  * Igbinary
  *
- * Igbinary is a compact binary format that serves as a drop-in replacement for the native PHP
- * serializer.
+ * Igbinary is a compact binary format that serves as a drop-in replacement for the native PHP serializer.
  *
  * @category    Machine Learning
  * @package     Rubix/ML
@@ -39,11 +38,18 @@ class Igbinary implements Serializer
      * @internal
      *
      * @param \Rubix\ML\Persistable $persistable
+     * @throws \Rubix\ML\Exceptions\RuntimeException
      * @return \Rubix\ML\Encoding
      */
     public function serialize(Persistable $persistable) : Encoding
     {
-        return new Encoding(igbinary_serialize($persistable) ?: '');
+        $data = igbinary_serialize($persistable);
+
+        if (!$data) {
+            throw new RuntimeException('Could not serialize data.');
+        }
+
+        return new Encoding($data);
     }
 
     /**
@@ -52,25 +58,19 @@ class Igbinary implements Serializer
      * @internal
      *
      * @param \Rubix\ML\Encoding $encoding
+     * @throws \Rubix\ML\Exceptions\RuntimeException
      * @return \Rubix\ML\Persistable
      */
     public function unserialize(Encoding $encoding) : Persistable
     {
-        $persistable = igbinary_unserialize((string) $encoding);
-
-        if (is_null($persistable)) {
-            throw new RuntimeException('Cannot read encoding, wrong'
-                . ' format or corrupted data.');
-        }
+        $persistable = igbinary_unserialize($encoding);
 
         if (!is_object($persistable)) {
-            throw new RuntimeException('Unserialized encoding must'
-                . ' be an object.');
+            throw new RuntimeException('Unserialized data must be an object.');
         }
 
         if ($persistable instanceof __PHP_Incomplete_Class) {
-            throw new RuntimeException('Missing class definition'
-                . ' for unserialized object.');
+            throw new RuntimeException('Missing class for object data.');
         }
 
         if (!$persistable instanceof Persistable) {