+++ src/Transformers/IntervalDiscretizer.php
@@ -6,6 +6,7 @@ use Tensor\Vector;
 use Rubix\ML\DataType;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Dataset;
+use Rubix\ML\Other\Traits\TracksRevisions;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
 use Rubix\ML\Exceptions\InvalidArgumentException;
 use Rubix\ML\Exceptions\RuntimeException;
@@ -26,6 +27,8 @@ use function is_null;
  */
 class IntervalDiscretizer implements Transformer, Stateful, Persistable
 {
+    use TracksRevisions;
+
     /**
      * The value of the starting category for each feature column.
      *