+++ src/Transformers/LinearDiscriminantAnalysis.php
@@ -7,6 +7,7 @@ use Rubix\ML\DataType;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Datasets\Dataset;
+use Rubix\ML\Other\Traits\TracksRevisions;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
 use Rubix\ML\Exceptions\InvalidArgumentException;
 use Rubix\ML\Exceptions\RuntimeException;
@@ -29,6 +30,8 @@ use const Rubix\ML\EPSILON;
  */
 class LinearDiscriminantAnalysis implements Transformer, Stateful, Persistable
 {
+    use TracksRevisions;
+
     /**
      * The target number of dimensions to project onto.
      *