+++ src/Transformers/MinMaxNormalizer.php
@@ -5,6 +5,7 @@ namespace Rubix\ML\Transformers;
 use Rubix\ML\DataType;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Dataset;
+use Rubix\ML\Other\Traits\TracksRevisions;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
 use Rubix\ML\Exceptions\InvalidArgumentException;
 use Rubix\ML\Exceptions\RuntimeException;
@@ -23,6 +24,8 @@ use const Rubix\ML\EPSILON;
  */
 class MinMaxNormalizer implements Transformer, Stateful, Elastic, Persistable
 {
+    use TracksRevisions;
+
     /**
      * The minimum value of the transformed features.
      *