+++ src/Transformers/TfIdfTransformer.php
@@ -5,6 +5,7 @@ namespace Rubix\ML\Transformers;
 use Rubix\ML\DataType;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Dataset;
+use Rubix\ML\Other\Traits\TracksRevisions;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
 use Rubix\ML\Exceptions\InvalidArgumentException;
 use Rubix\ML\Exceptions\RuntimeException;
@@ -32,6 +33,8 @@ use function is_null;
  */
 class TfIdfTransformer implements Transformer, Stateful, Elastic, Persistable
 {
+    use TracksRevisions;
+
     /**
      * The amount of additive (Laplace) smoothing to add to the inverse document
      * frequencies (IDFs).